requirements.txt
Flask
requests==2.32.3
pycryptodome==3.20.0
protobuf==6.30.0
google==3.0.0
Werkzeug==2.0.3
flask-cors
httpx
protobuf
cachetools
